pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
    "trafilatura>=2.0.0",
]
